# files/E-Commerce OpenAPI Specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E-Commerce API",
    "description": "\nWelcome to the Mock E-Commerce API! 🚀\n\nThis API provides a comprehensive set of functionalities for managing your e-commerce platform.\nKey features include:\n\n- **CRUD**\n\t- Create, Read, Update, and Delete endpoints.\n- **Search**\n\t- Find specific information with parameters and pagination.\n- **Auth**\n\t- Verify user/admin identity.\n\t- Secure with Access and Refresh tokens.\n- **Permission**\n\t- Assign roles with specific permissions.\n\t- Different access levels for User/Admin.\n- **Validation**\n\t- Ensure accurate and secure input data.\n\nFor any inquiries, please contact sshah@watermelon.us\n",
    "contact": {
      "name": "Sahil Shah",
      "url": "https://github.com/WM-SahilShah"
    },
    "version": "2.0.4"
  },
  "paths": {
...
    "/me/": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Get My Info #",
        "description": "Retrieve detailed information about the authenticated user. Requires valid authorization credentials.",
        "operationId": "get_my_info_me__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "Edit My Info #",
        "description": "Update user information, such as email, password, or other personal details. Requires valid authorization credentials.",
        "operationId": "edit_my_info_me__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Account"
        ],
        "summary": "Delete My Info #",
        "description": "Delete the current authenticated user's account permanently. Requires valid authorization credentials.",
        "operationId": "remove_my_account_me__delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get All Categories",
        "description": "This endpoint retrieves a paginated list of all categories with an optional search parameter to filter by category name.",
        "operationId": "get_all_categories_categories__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (Required)",
              "default": "<integer>*",
              "title": "Page"
            },
            "description": "Page number (Required)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page (Required)",
              "default": "<integer>*",
              "title": "Limit"
            },
            "description": "Items per page (Required)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Search based name of categories",
              "default": "<string>",
              "title": "Search"
            },
            "description": "Search based name of categories"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Create New Category ##",
        "description": "This endpoint allows an admin to create a new category.",
        "operationId": "create_category_categories__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{category_id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get Specific Category",
        "description": "This endpoint retrieves a specific category by its ID.",
        "operationId": "get_category_categories__category_id__get",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Update Existing Category ##",
        "description": "This endpoint allows an admin to update an existing category by its ID.",
        "operationId": "update_category_categories__category_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete Existing Category ##",
        "description": "This endpoint allows an admin to delete a category by its ID.",
        "operationId": "delete_category_categories__category_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOutDelete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountBase": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Unique identifier for the user or account.",
            "default": "<integer>"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "title": "Role",
            "description": "Role of the user (user or admin).",
            "default": "<string = `user`/`admin`"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Indicates whether the account is active.",
            "default": "<boolean>"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the account was created.",
            "default": "<datetime obj/ISO 8601 string>"
          },
          "carts": {
            "items": {
              "$ref": "#/components/schemas/CartBase"
            },
            "type": "array",
            "title": "Carts",
            "description": "List of carts associated with the account."
          },
          "username": {
            "type": "string",
            "minLength": 1,
            "title": "Username",
            "description": "Username of the account holder.",
            "default": "<string>"
          },
          "full_name": {
            "type": "string",
            "minLength": 1,
            "title": "Full Name",
            "description": "Full name of the account holder.",
            "default": "<string>"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email address of the account holder (validated).",
            "default": "email@example.com*"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "is_active",
          "created_at",
          "carts",
          "username",
          "full_name",
          "email"
        ],
        "title": "AccountBase",
        "description": "Represents details of an account.\n\nAttributes:\n- `id` (int): Unique identifier for the user or account.\n- `role` (str): Role of the user, either 'user' or 'admin'.\n- `is_active` (bool): Indicates whether the account is active.\n- `created_at` (datetime): Timestamp of account creation  in ISO 8601 format.\n- `carts` (List[CartBase]): List of carts associated with the account.\n- `username` (str): Username of the account holder.\n- `full_name` (str): Full name of the account holder.\n- `email` (str): Email address of the account holder (validated)."
      },
      "AccountOut": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message."
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountBase"
              }
            ],
            "description": "Account details."
          }
        },
        "type": "object",
        "required": [
          "message",
          "data"
        ],
        "title": "AccountOut",
        "description": "Represents the output schema for account responses.\n\nAttributes:\n- `message` (str): Response message.\n- `data` (AccountBase): Account details, including:\n    - `id` (int): Unique identifier for the user or account.\n    - `role` (str): Role of the user, either 'user' or 'admin'.\n    - `is_active` (bool): Indicates whether the account is active.\n    - `created_at` (datetime): Timestamp of account creation in ISO 8601 format.\n    - `carts` (List[CartBase]): List of carts associated with the account.\n    - `username` (str): Username of the account holder.\n    - `full_name` (str): Full name of the account holder.\n    - `email` (str): Email address of the account holder (validated)."
      },
      "AccountUpdate": {
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "title": "Username",
            "description": "Username of the account holder.",
            "default": "<string>"
          },
          "full_name": {
            "type": "string",
            "minLength": 1,
            "title": "Full Name",
            "description": "Full name of the account holder.",
            "default": "<string>"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email address of the account holder (validated).",
            "default": "email@example.com*"
          }
        },
        "type": "object",
        "title": "AccountUpdate",
        "description": "Represents schema for updating an account.\n\nAttributes:\n- `username` (str): Username of the account holder.\n- `full_name` (str): Full name of the account holder.\n- `email` (str): Email address of the account holder (validated)."
      },
...
      "CartBase": {
        "properties": {
          "id": {
            "type": "integer",
            "maximum": 399,
            "minimum": 300,
            "title": "Id",
            "description": "Unique identifier for the cart.",
            "default": "<integer>"
          },
          "user_id": {
            "type": "integer",
            "maximum": 599,
            "minimum": 500,
            "title": "User Id",
            "description": "Unique identifier for the user.",
            "default": "<integer>"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the cart was created.",
            "default": "<datetime obj/ISO 8601 string>"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount",
            "description": "Total amount for the cart.",
            "default": "<float>"
          },
          "cart_items": {
            "items": {
              "$ref": "#/components/schemas/CartItemBase"
            },
            "type": "array",
            "title": "Cart Items",
            "description": "List of items in the cart."
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "created_at",
          "total_amount",
          "cart_items"
        ],
        "title": "CartBase",
        "description": "Represents basic cart details.\n\nAttributes:\n- `id` (int): Unique identifier for the cart.\n- `user_id` (int): Unique identifier for the user.\n- `created_at` (datetime): Timestamp when the cart was created.\n- `total_amount` (float): Total amount for the cart.\n- `cart_items` (List[CartItemBase]): List of items in the cart.\n    - `id` (int): Unique identifier for the cart item.\n    - `product_id` (int): Unique identifier for the product.\n    - `quantity` (int): Quantity of the product in the cart.\n    - `subtotal` (float): Subtotal amount for the cart item.\n    - `product` (ProductBase): Details of the product."
      },
...
      "CartItemBase": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Unique identifier for the cart item.",
            "default": "<integer>"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id",
            "description": "Unique identifier for the product.",
            "default": "<integer>"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "description": "Quantity of the product in the cart.",
            "default": "<integer>"
          },
          "subtotal": {
            "type": "number",
            "title": "Subtotal",
            "description": "Subtotal amount for the cart item.",
            "default": "<float>"
          },
          "product": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductBase"
              }
            ],
            "description": "Details of the product."
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "quantity",
          "subtotal",
          "product"
        ],
        "title": "CartItemBase",
        "description": "Represents a cart item.\n\nAttributes:\n- `id` (int): Unique identifier for the cart item.\n- `product_id` (int): Unique identifier for the product.\n- `quantity` (int): Quantity of the product in the cart.\n- `subtotal` (float): Subtotal amount for the cart item.\n- `product` (ProductBase): Details of the product."
      },
...
      "CategoriesOut": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message."
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/CategoryBase"
            },
            "type": "array",
            "title": "Data",
            "description": "List of category details."
          }
        },
        "type": "object",
        "required": [
          "message",
          "data"
        ],
        "title": "CategoriesOut",
        "description": "Represents the output schema for multiple categories.\n\nAttributes:\n- `message` (str): Response message.\n- `data` (List[CategoryBase]): List of category details.\n    - `id` (int): Unique identifier for the category.\n    - `name` (str): Name of the category."
      },
      "CategoryBase": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Name of the category. Must have at least 1 character.",
            "default": "<string>"
          },
          "id": {
            "type": "integer",
            "maximum": 199,
            "minimum": 100,
            "title": "Id",
            "description": "Unique identifier for the category.",
            "default": "<integer>"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "CategoryBase",
        "description": "Represents the details of a category.\n\nAttributes:\n- `id` (int): Unique identifier for the category.\n- `name` (str): Name of the category."
      },
      "CategoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Name of the category. Must have at least 1 character.",
            "default": "<string>"
          }
        },
        "type": "object",
        "title": "CategoryCreate",
        "description": "Represents the basic details for creating a category.\n\nAttributes:\n- `name` (str): Name of the category."
      },
      "CategoryOut": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message."
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryBase"
              }
            ],
            "description": "Details of the category."
          }
        },
        "type": "object",
        "required": [
          "message",
          "data"
        ],
        "title": "CategoryOut",
        "description": "Represents the output schema for a single category.\n\nAttributes:\n- `message` (str): Response message.\n- `data` (CategoryBase): Details of the category.\n    - `id` (int): Unique identifier for the category.\n    - `name` (str): Name of the category."
      },
      "CategoryOutDelete": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message."
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryBase"
              }
            ],
            "description": "Details of the category."
          }
        },
        "type": "object",
        "required": [
          "message",
          "data"
        ],
        "title": "CategoryOutDelete",
        "description": "Represents the output schema for category deletion.\n\nAttributes:\n- `message` (str): Response message.\n- `data` (CategoryBase): Details of the deleted category.\n    - `id` (int): Unique identifier for the category.\n    - `name` (str): Name of the category."
      },
      "CategoryUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Name of the category. Must have at least 1 character.",
            "default": "<string>"
          }
        },
        "type": "object",
        "title": "CategoryUpdate",
        "description": "Represents the schema for updating a category.\n\nAttributes:\n- `name` (str): Name of the category."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProductBase": {
        "properties": {
          "id": {
            "type": "integer",
            "maximum": 299,
            "minimum": 200,
            "title": "Id",
            "description": "Unique identifier for the product.",
            "default": "<integer>"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Title of the product.",
            "default": "<string>"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description",
            "description": "Description of the product.",
            "default": "<string>"
          },
          "price": {
            "type": "integer",
            "minimum": 0,
            "title": "Price",
            "description": "Price of the product (>=0).",
            "default": "<integer>"
          },
          "discount_percentage": {
            "type": "number",
            "maximum": 100,
            "minimum": 0,
            "title": "Discount Percentage",
            "description": "Discount percentage (0-100).",
            "default": "<float>"
          },
          "rating": {
            "type": "number",
            "maximum": 5,
            "minimum": 0,
            "title": "Rating",
            "description": "Rating of the product (0-5).",
            "default": "<float>"
          },
          "stock": {
            "type": "integer",
            "minimum": 0,
            "title": "Stock",
            "description": "Stock count (>=0).",
            "default": "<integer>"
          },
          "brand": {
            "type": "string",
            "minLength": 1,
            "title": "Brand",
            "description": "Brand of the product.",
            "default": "<string>"
          },
          "thumbnail": {
            "type": "string",
            "minLength": 1,
            "title": "Thumbnail",
            "description": "URL of the product thumbnail.",
            "default": "<url>"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Images",
            "description": "List of image URLs for the product.",
            "default": "<url>"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "description": "Whether the product is published or not.",
            "default": "<boolean>"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the product was created.",
            "default": "<datetime obj/ISO 8601 string>"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id",
            "description": "Unique identifier for the product's category.",
            "default": "<integer>"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "price",
          "discount_percentage",
          "rating",
          "stock",
          "brand",
          "thumbnail",
          "images",
          "is_published",
          "created_at",
          "category_id"
        ],
        "title": "ProductBase",
        "description": "Represents basic details of a product.\n\nAttributes:\n- `id` (int): Unique identifier for the product.\n- `title` (str): Title of the product.\n- `description` (str): Description of the product.\n- `price` (int): Price of the product (>= 0).\n- `discount_percentage` (float): Discount percentage (0-100).\n- `rating` (float): Rating of the product (0-5).\n- `stock` (int): Stock count (>= 0).\n- `brand` (str): Brand of the product.\n- `thumbnail` (str): URL of the product thumbnail.\n- `images` (List[str]): List of image URLs for the product.\n- `is_published` (bool): Whether the product is published or not.\n- `created_at` (datetime): Timestamp when the product was created.\n- `category_id` (int): Unique identifier for the product's category."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}